requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
streamlit==1.28.0
pydantic==2.5.0
python-dotenv==1.0.0
httpx==0.25.0
pandas==2.1.3